+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add a reverse proxy to the docker-compose deployment and changed default ports to 80
 * Upgraded ska-db-migrations to 0.9.1
 * Added support to delete container tree hierarchies
+* Add Transactional Outbox Pattern with RabbitMQ
 * Add ODA trigger that inserts a dlm.data_item when an execution block is added.
 * Added dlm-archive stfp endpoint. 
 * Added pg_sphere to local db build.